move dataset loading functions to a new file
